pyproject.toml
[project]
name = "analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "clip-client>=0.8.3",
    "clip-server>=0.8.3",
    "gemimg>=0.3.2",
    "generativeai>=0.0.1",
    "google-cloud-aiplatform>=1.128.0",
    "google-genai>=1.52.0",
    "matplotlib>=3.10.7",
    "pandas>=2.3.3",
    "pca>=2.10.1",
    "pillow>=12.0.0",
    "seaborn>=0.13.2",
    "sentence-transformers>=5.1.2",
]
